requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
groq>=0.3.0
python-dotenv>=1.0.1
tiktoken>=0.5.0